refactor(app): the facade splits by subject, as the packages beside it already do
app.go was 1,973 lines and 67 declarations over seven subjects: the analyses
the frontend calls, saving a run and reading it back, exports and backups,
sign-in and the profile, the boot sequence, preferences, and a handful of
helpers. A reader looking for how a run is saved found it between Logout and
the mesh HTTP route.

The split is the one the package already uses. app_projects.go and
app_whiteboards.go were written this way; the rest of the facade was not.

  app_analysis.go   the analyses asked for by name, and the mesh route
  app_runs.go       saveRun, one persist* per product, and LoadAnalysis
  app_storage.go    what leaves as a file: export, backup, restore
  app_session.go    registration, sign-in, the profile, preferences
  app.go            the struct, the boot, and the two handles

Nothing moves between packages and no function changes: internal/ was already
cut by subject -- analysis, store, pyenv, research -- and this is the facade
above it catching up. app.go is 313 lines and says at its head what it does not
hold, so the next reader does not look here for a method that is next door.

app_runs.go is 880 of the 1,660 that moved, which is the same shape the sidecar
had before its own split: one subject carrying most of the weight, with one
function per product mapping that product's payload onto a common row.

Verification: gofmt clean, go vet clean, go test -race clean across every
package. golangci-lint is not installed on this machine; CI runs it.
