+++ app.go
@@ -2,15 +2,9 @@ package main
 
 import (
 	"context"
-	"encoding/base64"
-	"encoding/json"
 	"errors"
-	"fmt"
-	"io"
-	"net/http"
 	"os"
 	"path/filepath"
-	"strconv"
 	"strings"
 	"sync"
 	"time"
@@ -18,12 +12,21 @@ import (
 	"geosense-infer/internal/analysis"
 	"geosense-infer/internal/geocode"
 	"geosense-infer/internal/pyenv"
-	"geosense-infer/internal/research"
 	"geosense-infer/internal/store"
 
-	"github.com/google/uuid"
 	wruntime "github.com/wailsapp/wails/v2/pkg/runtime"
 )
+
+// The application struct the frontend is bound to, and what it takes to get
+// one running: the boot sequence, the log the splash reads, and the two
+// handles -- the sidecar runner and the local store -- every method reaches
+// through.
+//
+// What it does NOT hold, and why it is worth saying: the methods the frontend
+// calls are in app_analysis.go, app_runs.go, app_storage.go, app_session.go,
+// app_projects.go and app_whiteboards.go. This file held all of them, 67
+// declarations over seven subjects, and a reader looking for how a run is
+// saved had to find it among sign-in and window management.
 
 // App is the application struct bound to the frontend.
 type App struct {
@@ -300,800 +303,6 @@ func (a *App) probeSidecar(ctx context.Context) {
 	wruntime.EventsEmit(ctx, "boot:ready", ok)
 }
 
-// ListEmbeddedAreas returns the embedded study areas (A/B/C).
-func (a *App) ListEmbeddedAreas() []analysis.Area {
-	runner := a.currentRunner()
-	if runner == nil {
-		return []analysis.Area{}
-	}
-	return runner.ListAreas()
-}
-
-// Predict runs the inference sidecar for the given request.
-func (a *App) Predict(req analysis.PredictRequest) (*analysis.PredictResult, error) {
-	runner := a.currentRunner()
-	if runner == nil {
-		return nil, errors.New("runner not initialized")
-	}
-	res, err := runner.Predict(a.ctx, req)
-	if err != nil {
-		return nil, err
-	}
-	// Set after persisting, so the stored copy -- marshalled inside -- does not
-	// carry a run's own id inside its own row. The frontend needs it to attach
-	// compositions made while this run is on screen.
-	res.RunID = a.persistRunIfLoggedIn(req, res)
-	return res, nil
-}
-
-// AnalyzeLULC runs descriptive MapBiomas land-cover / land-use analysis
-// without Sentinel imagery. Embedded areas use local TIFFs; custom AOIs in
-// Brazil fetch a MapBiomas Collection 10 COG window on demand.
-func (a *App) AnalyzeLULC(req analysis.LULCRequest) (*analysis.LULCAnalysis, error) {
-	runner := a.currentRunner()
-	if runner == nil {
-		return nil, errors.New("runner not initialized")
-	}
-	return runner.AnalyzeLULC(a.ctx, req)
-}
-
-// ListDataCube inventories Sentinel-2 L2A scenes for the AOI (before Classify).
-func (a *App) ListDataCube(req analysis.DataCubeRequest) (*analysis.DataCubeResult, error) {
-	runner := a.currentRunner()
-	if runner == nil {
-		return nil, errors.New("runner not initialized")
-	}
-	return runner.ListDataCube(a.ctx, req)
-}
-
-// RenderComposite builds an RGB / false-color or spectral-index overlay for one scene.
-func (a *App) RenderComposite(req analysis.CompositeRequest) (*analysis.CompositeResult, error) {
-	runner := a.currentRunner()
-	if runner == nil {
-		return nil, errors.New("runner not initialized")
-	}
-	return runner.RenderComposite(a.ctx, req)
-}
-
-// AnalyzeWater maps surface water over a period from spectral water indices.
-// Descriptive: a thresholded index, with no model and no trained legend.
-func (a *App) AnalyzeWater(req analysis.WaterRequest) (*analysis.WaterAnalysis, error) {
-	runner := a.currentRunner()
-	if runner == nil {
-		return nil, errors.New("runner not initialized")
-	}
-	res, err := runner.AnalyzeWater(a.ctx, req)
-	if err != nil {
-		return nil, err
-	}
-	a.persistWaterRun(req, res)
-	return res, nil
-}
-
-/*
-savedRun is what one product contributes to a run row: the parts the writer
-below cannot know.
-
-Six products persist a run, and apart from these fields the path is one
-sequence written six times. That is how it drifted: the "run-" prefix, the
-trimmed project id and the AoiID column each had to be added in every copy,
-and the thumbnail column the classification path fills never reached any of
-the others.
-*/
-type savedRun struct {
-	// The discriminator every reader branches on, and what produced the
-	// numbers -- for the descriptive products a data source, not a model.
-	kind      string
-	modelKind string
-
-	// Where the run was made. The polygon wins; the area id is recorded as a
-	// reference when the request named a catalogued area instead of drawing
-	// a shape.
-	areaID  string
-	polygon *analysis.GeoJSONGeometry
-
-	// What the run is called. aoiLabel names the ground and arrives already
-	// resolved, because the summary the caller built carries the same string;
-	// runLabel is what the frontend asked for, and is minted from aoiLabel
-	// when it asked for nothing.
-	aoiLabel string
-	runLabel string
-
-	projectID string
-	// Which catalogued area this run is OF. The polygon says where it was
-	// made; the board needs the area to keep a drawing and its runs as one
-	// subject.
-	aoiID string
-
-	periodStart string
-	periodEnd   string
-	nDates      int
-
-	// The figures the saved-run list prints. This is the part that is really
-	// different between products, which is why it is built at the call site.
-	summary map[string]any
-
-	/*
-		result yields the payload to store, given the run's asset directory
-		and that directory's path relative to the data dir.
-
-		A function rather than a value because the assets have to be written
-		out, and their data URIs cleared from the copy being stored, before
-		there is anything to marshal: a result JSON still carrying its base64
-		would put a second copy of every overlay inside the database.
-	*/
-	result func(assetsDir, assetsRel string) any
-
-	// The file in that directory the run list paints its thumbnail from, for
-	// the products that write one.
-	overlayFile string
-}
-
-/*
-saveRun writes one run row and returns its id, empty when nothing was written.
-
-Best effort from end to end: failing to record a run must not discard the
-result the user is looking at, so every error along the way is dropped rather
-than reported.
-*/
-func (a *App) saveRun(run savedRun) string {
-	a.mu.RLock()
-	user := a.currentUser
-	st := a.store
-	a.mu.RUnlock()
-	if st == nil {
-		return ""
-	}
-	userID := store.LocalUserID
-	if user != nil {
-		userID = user.ID
-	}
-
-	runID := uuid.NewString()
-	assetsRel := filepath.Join("runs", runID)
-	assetsDir := st.RunsDir(runID)
-	// Each of the three below used to discard its error, and each failure
-	// produced the same shape of wreckage: a row that says a run was recorded,
-	// pointing at something that is not there.
-	//
-	// Without the directory, every asset write inside run.result fails and the
-	// row still carries assets_relpath and overlay_relpath into a path that
-	// does not exist. Without the marshal, result_json is written empty and
-	// LoadAnalysis returns a run with nothing in it -- and json.Marshal is not
-	// hypothetical here: it refuses NaN and Inf, which a sidecar payload with
-	// a missing float reaches the Go side carrying.
-	//
-	// Not recording the run at all is the better failure. The result still
-	// reaches the user, which is what best effort meant; only the claim to
-	// have saved it is withdrawn.
-	if err := os.MkdirAll(assetsDir, 0o700); err != nil {
-		return ""
-	}
-	resultBytes, err := json.Marshal(run.result(assetsDir, assetsRel))
-	if err != nil {
-		return ""
-	}
-	summaryBytes, err := json.Marshal(run.summary)
-	if err != nil {
-		return ""
-	}
-
-	poly := ""
-	if run.polygon != nil {
-		if b, err := json.Marshal(run.polygon); err == nil {
-			poly = string(b)
-		}
-	} else if run.areaID != "" {
-		poly = fmt.Sprintf(`{"area_id":%q}`, run.areaID)
-	}
-
-	runLabel := strings.TrimSpace(run.runLabel)
-	if runLabel == "" {
-		runLabel = makeRunLabel(run.aoiLabel)
-	}
-	// What makeRunLabel mints is prefixed already; a label that came in from
-	// the frontend need not be, and readers group runs on that prefix.
-	if !strings.HasPrefix(strings.ToLower(runLabel), "run-") {
-		runLabel = "run-" + runLabel
-	}
-
-	overlayRel := ""
-	if run.overlayFile != "" {
-		overlayRel = filepath.Join(assetsRel, run.overlayFile)
-	}
-
-	if _, err := st.SaveRun(store.InferenceRun{
-		ID:             runID,
-		UserID:         userID,
-		Kind:           run.kind,
-		ModelKind:      run.modelKind,
-		PeriodStart:    run.periodStart,
-		PeriodEnd:      run.periodEnd,
-		PolygonGeoJSON: poly,
-		Status:         "ok",
-		SummaryJSON:    string(summaryBytes),
-		ResultJSON:     string(resultBytes),
-		OverlayRelPath: overlayRel,
-		AssetsRelPath:  assetsRel,
-		NDates:         run.nDates,
-		Label:          runLabel,
-		ProjectID:      strings.TrimSpace(run.projectID),
-		AoiID:          strings.TrimSpace(run.aoiID),
-	}); err != nil {
-		// Best effort means the failure is not reported, not that it is
-		// reported as a success. The caller hands this id to the frontend as
-		// the run now on screen, and an id no row answers to is an entry that
-		// opens empty.
-		return ""
-	}
-	return runID
-}
-
-/*
-aoiLabel is what a run calls the ground it covers, taken from the first
-candidate the request actually filled in.
-
-Resolved in one place because the label reaches two: the run's own name and its
-summary. Resolved twice it can differ between them, and a run named for one
-area whose summary names another reads as two runs.
-*/
-func aoiLabel(candidates ...string) string {
-	for _, c := range candidates {
-		if s := strings.TrimSpace(c); s != "" {
-			return s
-		}
-	}
-	return "Custom AOI"
-}
-
-// persistWaterRun saves a surface-water run so it survives the session and is
-// listed, opened and exported like a classification.
-func (a *App) persistWaterRun(req analysis.WaterRequest, res *analysis.WaterAnalysis) {
-	if res == nil {
-		return
-	}
-	label := aoiLabel(req.Label)
-	a.saveRun(savedRun{
-		kind: store.RunKindWater,
-		// No model produced this: the index name carries the method instead.
-		modelKind:   res.Index,
-		areaID:      req.AreaID,
-		polygon:     req.PolygonGeoJSON,
-		aoiLabel:    label,
-		runLabel:    req.RunLabel,
-		projectID:   req.ProjectID,
-		aoiID:       req.AoiID,
-		periodStart: req.Start,
-		periodEnd:   req.End,
-		nDates:      res.NDates,
-		summary: map[string]any{
-			"water_index":             res.Index,
-			"n_dates":                 res.NDates,
-			"date_range":              res.DateRange,
-			"peak_date":               res.PeakDate,
-			"peak_water_fraction_pct": res.PeakWaterPct,
-			"ephemeral_area_ha":       res.EphemeralAreaHa,
-			"persistent_area_ha":      res.PersistentAreaHa,
-			"aoi_area_ha":             res.AOIAreaHa,
-			"aoi_label":               label,
-		},
-		result: func(assetsDir, _ string) any {
-			_ = store.WriteDataURIFile(
-				res.OccurrenceURI, filepath.Join(assetsDir, "water_occurrence.png"))
-			// Store without the bulky data URI; the asset is restored on load.
-			stored := *res
-			stored.OccurrenceURI = ""
-			return stored
-		},
-	})
-}
-
-// AnalyzeSolar computes the solar resource and photovoltaic yield at the AOI.
-func (a *App) AnalyzeSolar(req analysis.SolarRequest) (*analysis.SolarAnalysis, error) {
-	runner := a.currentRunner()
-	if runner == nil {
-		return nil, errors.New("runner not initialized")
-	}
-	res, err := runner.AnalyzeSolar(a.ctx, req)
-	if err != nil {
-		return nil, err
-	}
-	a.persistSolarRun(req, res)
-	return res, nil
-}
-
-// AnalyzeDomainShift compares two cached domain fingerprints for shift diagnosis.
-func (a *App) AnalyzeDomainShift(req analysis.DomainShiftRequest) (*analysis.DomainShiftReport, error) {
-	runner := a.currentRunner()
-	if runner == nil {
-		return nil, errors.New("runner not initialized")
-	}
-	return runner.AnalyzeDomainShift(a.ctx, req)
-}
-
-// AnalyzeDomainShiftCohort measures one source AOI against every target at once.
-func (a *App) AnalyzeDomainShiftCohort(
-	req analysis.DomainShiftCohortRequest,
-) (*analysis.DomainShiftCohort, error) {
-	runner := a.currentRunner()
-	if runner == nil {
-		return nil, errors.New("runner not initialized")
-	}
-	return runner.AnalyzeDomainShiftCohort(a.ctx, req)
-}
-
-// BuildCanopyField returns the leaf-area-density field of one orchard module,
-// together with the transmittances the GLSL march has to reproduce.
-//
-// Not persisted as a run: the field is a function of its parameters and costs
-// under a second to rebuild, so storing it would keep a copy that the next
-// change of spacing invalidates. The analyses that do get saved are the ones
-// carrying a satellite acquisition nobody can reproduce on demand.
-func (a *App) BuildCanopyField(req analysis.CanopyFieldRequest) (*analysis.CanopyField, error) {
-	runner := a.currentRunner()
-	if runner == nil {
-		return nil, errors.New("runner not initialized")
-	}
-	return runner.BuildCanopyField(a.ctx, req)
-}
-
-// BuildCanopyFromAOI reads an AOI's own vegetation-index series as a canopy:
-// LAI by date, the Helios age that carries it, and -- given a location -- what
-// that canopy intercepts under the sun the cell actually received.
-//
-// Not persisted, for the reason the other two canopy calls give: it is a
-// function of a saved run plus a sowing, and both are already recorded.
-func (a *App) BuildCanopyFromAOI(req analysis.CanopyFromAOIRequest) (*analysis.CanopyFromAOI, error) {
-	runner := a.currentRunner()
-	if runner == nil {
-		return nil, errors.New("runner not initialized")
-	}
-	return runner.BuildCanopyFromAOI(a.ctx, req)
-}
-
-// BuildCanopyMesh grows a stand of plants and returns it as glTF, for a reader
-// who wants to see the canopy rather than a density that stands for it.
-//
-// Not persisted, for the reason BuildCanopyField gives, and for one more: the
-// stand is deterministic in its seed, so the parameters are a smaller and more
-// durable record of it than the megabytes of triangles they produce.
-func (a *App) BuildCanopyMesh(req analysis.CanopyMeshRequest) (*analysis.CanopyMesh, error) {
-	runner := a.currentRunner()
-	if runner == nil {
-		return nil, errors.New("runner not initialized")
-	}
-	mesh, err := runner.BuildCanopyMesh(a.ctx, req)
-	if err != nil {
-		return nil, err
-	}
-
-	/*
-		The bytes are held here and the reply carries a URL instead.
-
-		Returning them would put a base64 string of the whole mesh through the
-		Wails bridge, which marshals every bound result to JSON. On WKWebView
-		that is where "Maximum call stack size exceeded" is thrown -- inside the
-		bridge, before any application JavaScript runs, which is why it survived
-		being verified everywhere outside the webview.
-
-		The id changes per build so the webview cannot serve a previous stand
-		from cache, and each build is held under its own id rather than
-		replacing the last -- see the field's comment for the race that made a
-		single slot wrong.
-	*/
-	id := uuid.NewString()
-
-	a.meshMu.Lock()
-	if a.meshes == nil {
-		a.meshes = make(map[string][]byte)
-	}
-	a.meshes[id] = mesh.Data
-	a.meshOrder = append(a.meshOrder, id)
-	for len(a.meshOrder) > maxHeldMeshes {
-		delete(a.meshes, a.meshOrder[0])
-		a.meshOrder = a.meshOrder[1:]
-	}
-	a.meshMu.Unlock()
-
-	mesh.Data = nil
-	mesh.URL = meshURLPrefix + id
-	return mesh, nil
-}
-
-// The path the grown stand is served from. A prefix rather than a fixed name
-// because the id changes per build, which is what keeps the webview from
-// answering a fetch out of its cache with the previous canopy.
-const meshURLPrefix = "/canopy-mesh/"
-
-/*
-meshMiddleware serves the last grown stand as bytes.
-
-AssetServer middleware rather than its Handler, and the distinction is the whole
-reason this works: Handler is consulted only when Assets reports the file
-missing, and a single-page front end answers any unknown path with index.html
-instead. A mesh request therefore came back as HTML and the loader reported
-"Unrecognized token '<'". Middleware sits ahead of Assets, so this decides its
-own route and passes everything else through untouched.
-*/
-func (a *App) meshMiddleware(next http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		if !strings.HasPrefix(r.URL.Path, meshURLPrefix) {
-			next.ServeHTTP(w, r)
-			return
-		}
-		id := strings.TrimPrefix(r.URL.Path, meshURLPrefix)
-
-		a.meshMu.RLock()
-		data, held := a.meshes[id]
-		a.meshMu.RUnlock()
-
-		// An id nobody is holding is one that has aged out, or one that was
-		// never issued. Either way this must answer rather than fall through:
-		// the asset server behind it replies to unknown paths with index.html,
-		// and a loader handed HTML reports "Unrecognized token '<'".
-		if id == "" || !held || len(data) == 0 {
-			http.NotFound(w, r)
-			return
-		}
-		w.Header().Set("Content-Type", "model/gltf-binary")
-		w.Header().Set("Content-Length", strconv.Itoa(len(data)))
-		// The id is unique per build, so the bytes behind a URL never change.
-		w.Header().Set("Cache-Control", "no-store")
-		w.WriteHeader(http.StatusOK)
-		_, _ = w.Write(data)
-	})
-}
-
-// persistSolarRun saves a solar resource run so it survives the session and is
-// listed, opened and exported like the other analyses.
-func (a *App) persistSolarRun(req analysis.SolarRequest, res *analysis.SolarAnalysis) {
-	if res == nil {
-		return
-	}
-	label := aoiLabel(req.Label)
-	a.saveRun(savedRun{
-		kind: store.RunKindSolar,
-		// No model produced this; the source is the method that did.
-		modelKind: "NASA POWER",
-		areaID:    req.AreaID,
-		polygon:   req.PolygonGeoJSON,
-		aoiLabel:  label,
-		runLabel:  req.RunLabel,
-		projectID: req.ProjectID,
-		aoiID:     req.AoiID,
-		nDates:    res.Resource.NYears,
-		summary: map[string]any{
-			"ghi_annual_kwh_m2":       res.Resource.GHIAnnualKWhM2,
-			"optimal_tilt_deg":        res.Geometry.OptimalTiltDeg,
-			"specific_yield":          res.PV.SpecificYieldKWhKWpYear,
-			"performance_ratio":       res.PV.PerformanceRatio,
-			"performance_ratio_model": res.PV.PerformanceRatioModelled,
-			"n_years":                 res.Resource.NYears,
-			"aoi_label":               label,
-			"grid_note":               res.GridNote,
-		},
-		result: func(string, string) any { return res },
-	})
-}
-
-// AnalyzeSolarTerrain maps plane-of-array irradiation over the AOI terrain.
-func (a *App) AnalyzeSolarTerrain(req analysis.SolarTerrainRequest) (*analysis.SolarTerrainAnalysis, error) {
-	runner := a.currentRunner()
-	if runner == nil {
-		return nil, errors.New("runner not initialized")
-	}
-	res, err := runner.AnalyzeSolarTerrain(a.ctx, req)
-	if err != nil {
-		return nil, err
-	}
-	a.persistSolarRaster(req.AreaID, req.PolygonGeoJSON, req.Label, req.RunLabel,
-		req.ProjectID, req.AoiID, "solar_terrain", res.Season, res, res.OverlayURI, res.NDates())
-	return res, nil
-}
-
-// AnalyzeSolarSiting classifies the AOI for fixed-tilt photovoltaic siting.
-func (a *App) AnalyzeSolarSiting(req analysis.SolarSitingRequest) (*analysis.SolarSitingAnalysis, error) {
-	runner := a.currentRunner()
-	if runner == nil {
-		return nil, errors.New("runner not initialized")
-	}
-	res, err := runner.AnalyzeSolarSiting(a.ctx, req)
-	if err != nil {
-		return nil, err
-	}
-	a.persistSolarRaster(req.AreaID, req.PolygonGeoJSON, req.Label, req.RunLabel,
-		req.ProjectID, req.AoiID, "solar_siting", "siting", res, res.OverlayURI, 0)
-	return res, nil
-}
-
-// persistSolarRaster saves a solar map run and writes its overlay to disk, so
-// reopening the run puts the raster back rather than only its numbers.
-func (a *App) persistSolarRaster(
-	areaID string, poly *analysis.GeoJSONGeometry,
-	label, runLabel, projectID, aoiID, kindTag, variant string,
-	payload any, overlayURI string, nDates int,
-) {
-	if payload == nil {
-		return
-	}
-	l := aoiLabel(label)
-	a.saveRun(savedRun{
-		kind:      store.RunKindSolar,
-		modelKind: "NASA POWER",
-		areaID:    areaID,
-		polygon:   poly,
-		aoiLabel:  l,
-		runLabel:  runLabel,
-		projectID: projectID,
-		aoiID:     aoiID,
-		nDates:    nDates,
-		summary: map[string]any{
-			"solar_product": kindTag,
-			"variant":       variant,
-			"aoi_label":     l,
-		},
-		result: func(assetsDir, _ string) any {
-			_ = store.WriteDataURIFile(overlayURI, filepath.Join(assetsDir, kindTag+".png"))
-			return payload
-		},
-	})
-}
-
-// AnalyzeEnergyModel runs the photovoltaic energy model over the AOI.
-func (a *App) AnalyzeEnergyModel(req analysis.EnergyModelRequest) (*analysis.EnergyModelAnalysis, error) {
-	runner := a.currentRunner()
-	if runner == nil {
-		return nil, errors.New("runner not initialized")
-	}
-	res, err := runner.AnalyzeEnergyModel(a.ctx, req)
-	if err != nil {
-		return nil, err
-	}
-	a.persistEnergyModelRun(req, res)
-	return res, nil
-}
-
-// persistEnergyModelRun saves an energy model run so it survives the session
-// and is listed, opened and exported like the other analyses.
-//
-// Filed under RunKindSolar with solar_product "energy_model", which is the
-// discriminator the solar products already use. It is a solar product: same
-// radiation chain, same grid, same optimum.
-func (a *App) persistEnergyModelRun(req analysis.EnergyModelRequest, res *analysis.EnergyModelAnalysis) {
-	if res == nil {
-		return
-	}
-	label := aoiLabel(req.Label)
-	a.saveRun(savedRun{
-		kind: store.RunKindSolar,
-		// No model produced this; the source is the method that did.
-		modelKind: "NASA POWER",
-		areaID:    req.AreaID,
-		polygon:   req.PolygonGeoJSON,
-		aoiLabel:  label,
-		runLabel:  req.RunLabel,
-		projectID: req.ProjectID,
-		aoiID:     req.AoiID,
-		nDates:    res.NDates(),
-		// ghi_annual_kwh_m2, optimal_tilt_deg and specific_yield are the keys
-		// the saved-run list already reads for a solar row, so the row says
-		// something without a client change. Everything else states the basis,
-		// because a yield without its performance ratio and reporting basis is
-		// not a figure.
-		summary: map[string]any{
-			"solar_product":            "energy_model",
-			"variant":                  res.ReportingBasis,
-			"ghi_annual_kwh_m2":        res.LossWaterfall.Base.GHIClimatologyKWhM2Year,
-			"optimal_tilt_deg":         res.Geometry.OptimalTiltDeg,
-			"specific_yield":           res.LossWaterfall.Delivered.AppliedKWhKWpYear,
-			"performance_ratio":        res.PerformanceRatio.Applied,
-			"performance_ratio_source": res.PerformanceRatio.AppliedSource,
-			"performance_ratio_model":  res.PerformanceRatio.Modelled,
-			"reporting_basis":          res.ReportingBasis,
-			"capacity_density_basis":   res.CapacityDensity.Basis,
-			"suitable_area_ha":         res.Plant.Suitable.AreaHa,
-			"suitable_capacity_dc_mw":  res.Plant.Suitable.CapacityDCMW,
-			"n_years":                  res.HourlyYears,
-			"aoi_label":                label,
-			"grid_note":                res.GridNote,
-		},
-		result: func(string, string) any { return res },
-	})
-}
-
-// AnalyzeWind screens the wind resource at the AOI.
-func (a *App) AnalyzeWind(req analysis.WindRequest) (*analysis.WindAnalysis, error) {
-	runner := a.currentRunner()
-	if runner == nil {
-		return nil, errors.New("runner not initialized")
-	}
-	res, err := runner.AnalyzeWind(a.ctx, req)
-	if err != nil {
-		return nil, err
-	}
-	a.persistWindRun(req, res)
-	return res, nil
-}
-
-// persistWindRun saves a wind screening run under its own kind.
-func (a *App) persistWindRun(req analysis.WindRequest, res *analysis.WindAnalysis) {
-	if res == nil {
-		return
-	}
-	label := aoiLabel(req.Label)
-	a.saveRun(savedRun{
-		kind: store.RunKindWind,
-		// No model produced this; the source is the product that did.
-		modelKind: "NASA POWER MERRA-2",
-		areaID:    req.AreaID,
-		polygon:   req.PolygonGeoJSON,
-		aoiLabel:  label,
-		runLabel:  req.RunLabel,
-		projectID: req.ProjectID,
-		aoiID:     req.AoiID,
-		nDates:    res.NDates(),
-		// The qualifier and the check outcome travel with the capacity factor.
-		// A gross, unvalidated figure listed beside a benchmarked photovoltaic
-		// one reads as the same kind of number unless the row says otherwise.
-		summary: map[string]any{
-			"wind_hub_height_m":              res.HubHeightM,
-			"wind_mean_speed_50m_ms":         res.Measured.MeanSpeed50mMS,
-			"wind_hub_mean_speed_ms":         res.Hub.MeanSpeedMS,
-			"wind_gross_capacity_factor_pct": res.Hub.GrossCapacityFactorPct,
-			"wind_annual_energy_mwh":         res.Hub.GrossAnnualEnergyMWhPerTurbine,
-			"wind_turbine":                   res.Turbine.Name,
-			"wind_all_checks_passed":         res.DataQuality.AllChecksPassed,
-			"wind_flag_count":                len(res.DataQuality.Flags),
-			"record_window":                  res.RecordWindow,
-			"qualifier":                      res.Qualifier,
-			"aoi_label":                      label,
-			"grid_note":                      res.GridNote,
-		},
-		result: func(string, string) any { return res },
-	})
-}
-
-/*
-AnalyzeFlood measures the HAND flood extent over the AOI and how much of that
-extent the choice of DEM product decides rather than the terrain.
-
-There is no call that returns one mask. What comes back is the agreement count
-raster -- per cell, how many products call it flooded at the reference
-threshold -- with the pairwise envelope around it, because an extent shipped
-alone is a shape produced by a DEM the user never chose and is never shown.
-*/
-func (a *App) AnalyzeFlood(req analysis.FloodRequest) (*analysis.FloodAnalysis, error) {
-	runner := a.currentRunner()
-	if runner == nil {
-		return nil, errors.New("runner not initialized")
-	}
-	res, err := runner.AnalyzeFlood(a.ctx, req)
-	if err != nil {
-		return nil, err
-	}
-	a.persistFloodRun(req, res)
-	return res, nil
-}
-
-// persistFloodRun saves a flood envelope run under its own kind, with its two
-// rasters copied out of the sidecar's work directory and the bounds that place
-// the displayed one stored beside them.
-//
-// The copy is the point. AnalyzeFlood deliberately leaves the work directory in
-// place because the returned paths point into it, but that directory is under
-// the system temporary root and is removed on a schedule nobody here controls.
-// A run whose stored paths still named it would list, reopen, and hand the
-// reader a GeoTIFF path that resolves to nothing -- and the GeoTIFF is the
-// product, not an illustration of it.
-func (a *App) persistFloodRun(req analysis.FloodRequest, res *analysis.FloodAnalysis) {
-	if res == nil {
-		return
-	}
-	label := aoiLabel(req.Label)
-	// The envelope row at the reference threshold: the narrowest and widest
-	// pairwise agreement where the agreement raster was built. Absent when the
-	// reference threshold produced no defined pair, in which case the summary
-	// carries no range rather than a zero one.
-	var refRow *analysis.FloodEnvelopeRow
-	for i := range res.Envelope {
-		if res.Envelope[i].ThresholdM == res.ReferenceThresholdM {
-			refRow = &res.Envelope[i]
-			break
-		}
-	}
-	summary := map[string]any{
-		"flood_reference_threshold_m": res.ReferenceThresholdM,
-		"flood_drainage_km2":          res.DrainageKm2,
-		"flood_n_products":            len(res.Products),
-		"flood_products":              floodProductIDs(res.Products),
-		/*
-			The AOI area, beside the two areas measured inside it.
-
-			The agreement figures below are now taken over the AOI polygon
-			rather than over the buffered window the terrain chain ran on, and
-			a row listing a contested area with no denominator cannot be told
-			apart from one that was taken over the window -- the difference on
-			the recorded payload is 4.5 km2 against 37.5. Listed here so the
-			run row carries the ground its own numbers are of.
-		*/
-		"flood_aoi_area_km2":      res.AOI.AreaKm2,
-		"flood_unanimous_wet_km2": res.Agreement.UnanimousWetKm2,
-		"flood_contested_km2":     res.Agreement.ContestedKm2,
-		// Null when no product calls anything wet. Zero would read as four
-		// products agreeing on an extent, which is the opposite of the fact.
-		"flood_contested_frac_of_wet": res.Agreement.ContestedFracOfWet,
-		/*
-			The qualifier is a listed field and not a detail of the open card.
-
-			Every figure above is a measurement over TERRA's own DEM set and
-			none of them is a flood depth, an extent or a probability. A row
-			printing a contested area with nothing beside it is exactly the
-			reading the qualifier exists to prevent, and the run list is where
-			these figures are read most often and explained least.
-		*/
-		"qualifier": res.Qualifier,
-		"aoi_label": label,
-	}
-	if refRow != nil {
-		summary["flood_iou_min"] = refRow.IoUMin
-		summary["flood_iou_max"] = refRow.IoUMax
-	}
-	// The thumbnail is claimed only when there is a rendering to write. Named
-	// unconditionally, the row would carry an overlay path into a file the
-	// result closure never produced, which is the shape saveRun refuses to
-	// write a row for elsewhere.
-	overlay := ""
-	if res.AgreementURI != "" || res.AgreementPNG != "" {
-		overlay = floodAgreementPNG
-	}
-	a.saveRun(savedRun{
-		kind: store.RunKindFlood,
-		// No model produced this; the terrain index and the catalogue the DEMs
-		// were read from are the method.
-		modelKind: "HAND over Planetary Computer DEM",
-		areaID:    req.AreaID,
-		polygon:   req.PolygonGeoJSON,
-		aoiLabel:  label,
-		runLabel:  req.RunLabel,
-		projectID: req.ProjectID,
-		aoiID:     req.AoiID,
-		// Zero, and not unfilled: HAND is a static terrain index with no
-		// observation dates to count. See FloodAnalysis.NDates.
-		nDates:  res.NDates(),
-		summary: summary,
-		result: func(assetsDir, assetsRel string) any {
-			stored := *res
-			// The rendering, written from the data URI when there is one and
-			// from the sidecar's path when the URI could not be read.
-			pngSrc := res.AgreementURI
-			if pngSrc == "" {
-				pngSrc = res.AgreementPNG
-			}
-			stored.AgreementPNG = ""
-			if err := store.WriteDataURIFile(
-				pngSrc, filepath.Join(assetsDir, floodAgreementPNG),
-			); err == nil && pngSrc != "" {
-				stored.AgreementPNG = filepath.Join(assetsRel, floodAgreementPNG)
-			}
-			stored.AgreementTIF = ""
-			if strings.TrimSpace(res.AgreementTIF) != "" {
-				if err := store.WriteDataURIFile(
-					res.AgreementTIF, filepath.Join(assetsDir, floodAgreementTIF),
-				); err == nil {
-					stored.AgreementTIF = filepath.Join(assetsRel, floodAgreementTIF)
-				}
-			}
-			// The base64 stays out of the database; LoadAnalysis reads it back
-			// from the file written above.
-			stored.AgreementURI = ""
-			return stored
-		},
-		// The run list paints its thumbnail from the agreement raster, which is
-		// the one image this product has.
-		overlayFile: overlay,
-	})
-}
-
 // The two files a flood run keeps in its assets directory. Named in one place
 // because the persist path writes them and LoadAnalysis reads them, and a run
 // whose writer and reader disagree on a file name reopens without its raster.
@@ -1101,665 +310,6 @@ const (
 	floodAgreementPNG = "flood_agreement.png"
 	floodAgreementTIF = "flood_agreement.tif"
 )
-
-// floodProductIDs lists which DEM products the envelope was measured over, for
-// the run row. The envelope is a property of the set, so a range listed without
-// the set it spans is not attributable to anything.
-func floodProductIDs(products []analysis.FloodProduct) []string {
-	ids := make([]string, 0, len(products))
-	for _, p := range products {
-		ids = append(ids, p.ID)
-	}
-	return ids
-}
-
-// ExportClassification copies the classification GeoTIFF to a user-chosen path.
-func (a *App) ExportClassification(rasterPath string) (string, error) {
-	return a.ExportOverlayFile(rasterPath, "terra_classification.tif")
-}
-
-// ExportResearchPack writes a ZIP of tabular CSVs (+ AOI / optional GeoTIFF)
-// for use in an external training or study workspace.
-func (a *App) ExportResearchPack(meta analysis.ResearchExportMeta, result *analysis.PredictResult) (string, error) {
-	if result == nil {
-		return "", errors.New("no analysis result to export")
-	}
-	dataDir := ""
-	if st := a.currentStore(); st != nil {
-		dataDir = st.DataDir()
-	}
-	zipBytes, err := research.BuildResearchPackZIP(meta, result, dataDir)
-	if err != nil {
-		return "", err
-	}
-
-	dest, err := wruntime.SaveFileDialog(a.ctx, wruntime.SaveDialogOptions{
-		Title:           "Export research pack",
-		DefaultFilename: research.DefaultResearchPackFilename(meta.AoiLabel),
-		Filters: []wruntime.FileFilter{
-			{DisplayName: "ZIP archive", Pattern: "*.zip"},
-		},
-	})
-	if err != nil {
-		return "", err
-	}
-	if dest == "" {
-		return "", nil
-	}
-	if !strings.HasSuffix(strings.ToLower(dest), ".zip") {
-		dest += ".zip"
-	}
-	if err := os.WriteFile(dest, zipBytes, 0o644); err != nil {
-		return "", err
-	}
-	return dest, nil
-}
-
-/*
-ExportBackup writes the local database and its files to a ZIP the user places.
-
-Everything this application saves lives in one directory on one machine: there
-is no server and no account holding a second copy, so a reinstalled laptop
-takes every analysis and project with it. This is the only way out.
-
-Password hashes and session tokens are left out of the archive. A backup is a
-file people mail to themselves and attach to support threads, and any of those
-turns a stored credential into one that has left the machine. Restoring returns
-the analyses and projects and asks for a new password; the archive says so in
-its README, and so does the button that makes it.
-
-Returns the path written, or an empty string when the user cancels the dialog.
-*/
-func (a *App) ExportBackup() (string, error) {
-	st := a.currentStore()
-	if st == nil {
-		return "", errors.New("the local store is not open")
-	}
-
-	// Built before the dialog opens. A large history takes a moment to archive,
-	// and doing it after would leave the user looking at a chosen filename with
-	// nothing happening, unable to tell a slow write from a stuck one.
-	zipBytes, err := st.ExportBackup(a.GetAppVersion())
-	if err != nil {
-		return "", err
-	}
-
-	dest, err := wruntime.SaveFileDialog(a.ctx, wruntime.SaveDialogOptions{
-		Title:           "Export backup",
-		DefaultFilename: store.DefaultBackupFilename(time.Now()),
-		Filters: []wruntime.FileFilter{
-			{DisplayName: "ZIP archive", Pattern: "*.zip"},
-		},
-	})
-	if err != nil {
-		return "", err
-	}
-	if dest == "" {
-		return "", nil
-	}
-	if !strings.HasSuffix(strings.ToLower(dest), ".zip") {
-		dest += ".zip"
-	}
-	if err := os.WriteFile(dest, zipBytes, 0o600); err != nil {
-		return "", err
-	}
-	return dest, nil
-}
-
-// ChooseBackupArchive opens a file dialog and describes the archive picked,
-// without changing anything.
-//
-// Separate from RestoreBackup so the user is told what they are about to get
-// and what it will displace before it happens. A restore replaces everything;
-// an operation of that weight should not be one click from a file dialog.
-func (a *App) ChooseBackupArchive() (*store.RestorePreview, error) {
-	st := a.currentStore()
-	if st == nil {
-		return nil, errors.New("the local store is not open")
-	}
-
-	path, err := wruntime.OpenFileDialog(a.ctx, wruntime.OpenDialogOptions{
-		Title: "Choose a TERRA backup",
-		Filters: []wruntime.FileFilter{
-			{DisplayName: "ZIP archive", Pattern: "*.zip"},
-		},
-	})
-	if err != nil {
-		return nil, err
-	}
-	if path == "" {
-		return nil, nil // Cancelled.
-	}
-
-	preview, err := st.InspectBackup(path)
-	if err != nil {
-		return nil, err
-	}
-	preview.ArchivePath = path
-	return preview, nil
-}
-
-/*
-RestoreBackup replaces the local data with the archive at the given path.
-
-The path comes from ChooseBackupArchive rather than from the frontend picking a
-file itself, so the thing restored is the thing that was described.
-
-The store is closed, replaced and reopened here. The database file is swapped
-underneath the connection, and a connection held across that points at a file
-that no longer exists -- so every later query would fail in a way that looks
-like corruption rather than like a restore.
-*/
-func (a *App) RestoreBackup(archivePath string) (*store.RestoreResult, error) {
-	if strings.TrimSpace(archivePath) == "" {
-		return nil, errors.New("no backup was chosen")
-	}
-
-	st := a.currentStore()
-	if st == nil {
-		return nil, errors.New("the local store is not open")
-	}
-
-	// Checked again here, not only in the preview: the file may have been
-	// replaced between describing it and restoring it, and this is the last
-	// point at which refusing costs nothing.
-	preview, err := st.InspectBackup(archivePath)
-	if err != nil {
-		return nil, err
-	}
-	if preview.Problem != "" {
-		return nil, errors.New(preview.Problem)
-	}
-
-	// Dropped from the field before it is closed. Close leaves the pointer
-	// valid and its database shut, so between here and the reopen below --
-	// extraction, verification, a directory rename -- every concurrent binding
-	// call reached the driver and came back "sql: database is closed" instead
-	// of being told the store was unavailable.
-	a.mu.Lock()
-	a.store = nil
-	a.mu.Unlock()
-
-	// Closed before the swap. RestoreBackup renames the directory this
-	// connection's file lives in.
-	if err := st.Close(); err != nil {
-		return nil, fmt.Errorf("closing the local store: %w", err)
-	}
-
-	result, restoreErr := st.RestoreBackup(archivePath)
-
-	// Reopened whether or not the restore worked. A failed restore leaves the
-	// previous data in place, and leaving the application with a closed store
-	// would turn a recoverable failure into one that needs a relaunch.
-	reopened, openErr := store.Open()
-	if openErr != nil {
-		if restoreErr != nil {
-			return nil, restoreErr
-		}
-		return nil, fmt.Errorf("the data was restored but could not be reopened: %w", openErr)
-	}
-	a.mu.Lock()
-	a.store = reopened
-	// The session belonged to the replaced database. Restored accounts carry no
-	// password hash, so there is nothing to be signed in as.
-	a.currentUser = nil
-	a.sessionToken = ""
-	a.mu.Unlock()
-
-	if restoreErr != nil {
-		return nil, restoreErr
-	}
-	return result, nil
-}
-
-// InspectStorage reports what the local data is made of.
-//
-// Measured by walking the directory rather than inferred from the database:
-// the database records what was saved, not what is on disk, and this screen is
-// only worth having if it is believed when it says where the space went.
-func (a *App) InspectStorage() (*store.StorageReport, error) {
-	st := a.currentStore()
-	if st == nil {
-		return nil, errors.New("the local store is not open")
-	}
-	return st.InspectStorage()
-}
-
-// PurgeOrphanedRunAssets removes run files with no analysis pointing at them.
-//
-// The only deletion offered without naming what is being deleted, because
-// these are the only files nothing can reach. Anything else is removed by
-// removing the analysis it belongs to, which the user does deliberately.
-func (a *App) PurgeOrphanedRunAssets() (*store.PurgeResult, error) {
-	st := a.currentStore()
-	if st == nil {
-		return nil, errors.New("the local store is not open")
-	}
-	removed, freed, err := st.PurgeOrphanedRunAssets()
-	if err != nil {
-		return nil, err
-	}
-	return &store.PurgeResult{Removed: removed, FreedBytes: freed}, nil
-}
-
-// ExportOverlayFile saves an overlay asset via SaveFileDialog.
-// src may be a filesystem path or a data:image/png;base64,... URI.
-func (a *App) ExportOverlayFile(src string, defaultFilename string) (string, error) {
-	src = strings.TrimSpace(src)
-	if src == "" {
-		return "", errors.New("no overlay to export")
-	}
-	if strings.TrimSpace(defaultFilename) == "" {
-		defaultFilename = "terra_overlay.png"
-	}
-
-	ext := strings.ToLower(filepath.Ext(defaultFilename))
-	filters := []wruntime.FileFilter{
-		{DisplayName: "PNG", Pattern: "*.png"},
-		{DisplayName: "GeoTIFF", Pattern: "*.tif;*.tiff"},
-	}
-	switch ext {
-	case ".tif", ".tiff":
-		filters = []wruntime.FileFilter{
-			{DisplayName: "GeoTIFF", Pattern: "*.tif;*.tiff"},
-		}
-	case ".png":
-		filters = []wruntime.FileFilter{
-			{DisplayName: "PNG", Pattern: "*.png"},
-		}
-	}
-
-	dest, err := wruntime.SaveFileDialog(a.ctx, wruntime.SaveDialogOptions{
-		Title:           "Export overlay",
-		DefaultFilename: defaultFilename,
-		Filters:         filters,
-	})
-	if err != nil {
-		return "", err
-	}
-	if dest == "" {
-		return "", nil
-	}
-
-	if strings.HasPrefix(src, "data:") {
-		raw, err := decodeDataURI(src)
-		if err != nil {
-			return "", err
-		}
-		if err := os.WriteFile(dest, raw, 0o644); err != nil {
-			return "", err
-		}
-		return dest, nil
-	}
-
-	if _, err := os.Stat(src); err != nil {
-		return "", fmt.Errorf("overlay file not found (re-apply or re-run to regenerate): %s", src)
-	}
-	in, err := os.Open(src)
-	if err != nil {
-		return "", err
-	}
-	defer in.Close()
-	out, err := os.Create(dest)
-	if err != nil {
-		return "", err
-	}
-	if _, err := io.Copy(out, in); err != nil {
-		out.Close()
-		return "", err
-	}
-	// Closed here rather than deferred and dropped. A write can fail for the
-	// first time at Close -- a full disk or an exceeded quota is reported when
-	// the last buffered bytes are flushed -- and discarding that error handed
-	// back the path of a truncated file as an export that had worked.
-	if err := out.Close(); err != nil {
-		return "", err
-	}
-	return dest, nil
-}
-
-func decodeDataURI(uri string) ([]byte, error) {
-	const marker = "base64,"
-	i := strings.Index(uri, marker)
-	if i < 0 {
-		return nil, errors.New("unsupported data URI")
-	}
-	return base64.StdEncoding.DecodeString(uri[i+len(marker):])
-}
-
-// persistRunIfLoggedIn saves the run and returns its id, so the caller can tell
-// the frontend which run it is now looking at. Empty when nothing was saved.
-func (a *App) persistRunIfLoggedIn(req analysis.PredictRequest, res *analysis.PredictResult) string {
-	return a.persistAnalysis(req, res)
-}
-
-func (a *App) persistAnalysis(req analysis.PredictRequest, res *analysis.PredictResult) string {
-	if res == nil {
-		return ""
-	}
-	label := aoiLabel(req.Label, req.AreaID)
-	runID := a.saveRun(savedRun{
-		kind:        store.RunKindClassification,
-		modelKind:   req.ModelKind,
-		areaID:      req.AreaID,
-		polygon:     req.PolygonGeoJSON,
-		aoiLabel:    label,
-		runLabel:    req.RunLabel,
-		projectID:   req.ProjectID,
-		aoiID:       req.AoiID,
-		periodStart: req.Start,
-		periodEnd:   req.End,
-		nDates:      res.NDates,
-		summary: map[string]any{
-			"class_stats":     res.ClassStats,
-			"date_range":      res.DateRange,
-			"n_dates":         res.NDates,
-			"mean_confidence": res.MeanConfidence,
-			"area_id":         req.AreaID,
-			"aoi_label":       label,
-			"has_reference":   res.ReferenceURI != "",
-			"has_ndvi_mean":   res.NDVIMeanURI != "",
-			"has_true_color":  res.TrueColorURI != "",
-		},
-		// The only path that writes more than one asset, and the reason the
-		// stored payload is produced here rather than handed over: the raster
-		// records where it landed, and the five images have to be gone from
-		// the copy that reaches the database.
-		result: func(assetsDir, assetsRel string) any {
-			_ = store.WriteDataURIFile(res.OverlayURI, filepath.Join(assetsDir, "overlay.png"))
-			_ = store.WriteDataURIFile(res.ConfidenceURI, filepath.Join(assetsDir, "confidence.png"))
-			_ = store.WriteDataURIFile(res.NDVIMeanURI, filepath.Join(assetsDir, "ndvi_mean.png"))
-			_ = store.WriteDataURIFile(res.TrueColorURI, filepath.Join(assetsDir, "true_color.png"))
-			_ = store.WriteDataURIFile(res.ReferenceURI, filepath.Join(assetsDir, "reference.png"))
-			if res.LULC != nil && res.LULC.MapURI != "" {
-				_ = store.WriteDataURIFile(res.LULC.MapURI, filepath.Join(assetsDir, "lulc_map.png"))
-			}
-			rasterRel := ""
-			if strings.TrimSpace(res.RasterTIF) != "" {
-				dest := filepath.Join(assetsDir, "classification.tif")
-				if err := store.WriteDataURIFile(res.RasterTIF, dest); err == nil {
-					rasterRel = filepath.Join(assetsRel, "classification.tif")
-				}
-			}
-
-			// Persist result without bulky data URIs; assets restored on load.
-			stored := *res
-			stored.OverlayURI = ""
-			stored.ConfidenceURI = ""
-			stored.NDVIMeanURI = ""
-			stored.TrueColorURI = ""
-			stored.ReferenceURI = ""
-			if stored.LULC != nil {
-				lulcCopy := *stored.LULC
-				lulcCopy.MapURI = ""
-				lulcCopy.MapPNG = ""
-				stored.LULC = &lulcCopy
-			}
-			// Empty when the raster was absent or could not be written, so the
-			// stored path never points at a file that is not there.
-			stored.RasterTIF = rasterRel
-			return stored
-		},
-		overlayFile: "overlay.png",
-	})
-	if runID != "" && strings.TrimSpace(req.ProjectID) != "" {
-		if st := a.currentStore(); st != nil {
-			st.TouchProject(req.ProjectID)
-		}
-	}
-	return runID
-}
-
-// ListRuns returns recent inference runs (signed-in user, or local guest).
-func (a *App) ListRuns(limit int) ([]store.InferenceRun, error) {
-	st, err := a.requireStore()
-	if err != nil {
-		return nil, err
-	}
-	a.mu.RLock()
-	u := a.currentUser
-	a.mu.RUnlock()
-	userID := store.LocalUserID
-	if u != nil {
-		userID = u.ID
-	}
-	return st.ListRuns(userID, limit)
-}
-
-// RunActivity returns the number of runs per calendar day over a trailing
-// window, for the activity grid on the settings screen.
-//
-// Separate from ListRuns because that one caps at 100 rows and carries the full
-// result payload on each: a year of activity read through it would show empty
-// weeks that are not empty.
-func (a *App) RunActivity(days int) ([]store.ActivityDay, error) {
-	st, err := a.requireStore()
-	if err != nil {
-		return nil, err
-	}
-	a.mu.RLock()
-	u := a.currentUser
-	a.mu.RUnlock()
-	userID := store.LocalUserID
-	if u != nil {
-		userID = u.ID
-	}
-	return st.RunActivity(userID, days)
-}
-
-// LoadAnalysis restores a saved PredictResult (with image data URIs) by run id.
-func (a *App) LoadAnalysis(runID string) (*analysis.PredictResult, error) {
-	st, err := a.requireStore()
-	if err != nil {
-		return nil, err
-	}
-	a.mu.RLock()
-	u := a.currentUser
-	a.mu.RUnlock()
-	userID := store.LocalUserID
-	if u != nil {
-		userID = u.ID
-	}
-	run, err := st.GetRun(userID, runID)
-	if err != nil {
-		// Also try local bucket if signed-in user has no match (legacy local saves).
-		if u != nil {
-			run, err = st.GetRun(store.LocalUserID, runID)
-		}
-		if err != nil {
-			return nil, mapStoreErr(err)
-		}
-	}
-	assetsDir := st.RunsDir(run.ID)
-
-	// A water run stores a WaterAnalysis, not a PredictResult. It is returned
-	// attached to an otherwise empty result so the analysis view and the export
-	// see it through the same field a live run uses. The classification fields
-	// are deliberately left at zero: no classification was made, and filling
-	// n_dates here would make the page present one.
-	if run.Kind == store.RunKindSolar {
-		// One kind covers four products; the summary says which one was saved.
-		var meta struct {
-			Product string `json:"solar_product"`
-		}
-		_ = json.Unmarshal([]byte(run.SummaryJSON), &meta)
-		out := &analysis.PredictResult{}
-		switch meta.Product {
-		case "solar_terrain":
-			var t analysis.SolarTerrainAnalysis
-			_ = json.Unmarshal([]byte(run.ResultJSON), &t)
-			if uri, err := store.ReadFileDataURI(
-				filepath.Join(assetsDir, "solar_terrain.png"), "image/png",
-			); err == nil {
-				t.OverlayURI = uri
-			}
-			out.SolarTerrain = &t
-		case "solar_siting":
-			var st analysis.SolarSitingAnalysis
-			_ = json.Unmarshal([]byte(run.ResultJSON), &st)
-			if uri, err := store.ReadFileDataURI(
-				filepath.Join(assetsDir, "solar_siting.png"), "image/png",
-			); err == nil {
-				st.OverlayURI = uri
-			}
-			out.SolarSiting = &st
-		// "energy_advanced" is the tag this product was written under before it
-		// was renamed. It is read and never written. Runs saved under the old
-		// tag would otherwise fall to the default branch, which decodes an
-		// energy payload into a SolarAnalysis and reopens as an empty solar
-		// card with nothing raising an error; a rename that makes a saved run
-		// unopenable is a worse defect than the name it fixes. The real store
-		// held no such row when the rename was made, so nothing was migrated
-		// and no write path can produce the old tag again.
-		case "energy_model", "energy_advanced":
-			// No raster: the whole run is in result_json. Without this case the
-			// run saves and lists correctly and reopens as an empty solar card,
-			// with nothing raising an error.
-			var e analysis.EnergyModelAnalysis
-			if run.ResultJSON != "" && run.ResultJSON != "{}" {
-				_ = json.Unmarshal([]byte(run.ResultJSON), &e)
-			}
-			e.NormalizeNilSlices()
-			out.EnergyModel = &e
-		default:
-			var solar analysis.SolarAnalysis
-			if run.ResultJSON != "" && run.ResultJSON != "{}" {
-				_ = json.Unmarshal([]byte(run.ResultJSON), &solar)
-			}
-			out.Solar = &solar
-		}
-		out.RunID = run.ID
-		return out, nil
-	}
-
-	// A wind run stores a WindAnalysis and no raster. Returned attached to an
-	// otherwise empty result, the way a water run is, so the analysis view and
-	// the export see it through the field a live run uses.
-	if run.Kind == store.RunKindWind {
-		var wind analysis.WindAnalysis
-		if run.ResultJSON != "" && run.ResultJSON != "{}" {
-			_ = json.Unmarshal([]byte(run.ResultJSON), &wind)
-		}
-		wind.NormalizeNilSlices()
-		return &analysis.PredictResult{Wind: &wind}, nil
-	}
-
-	/*
-		A flood run stores a FloodAnalysis and two rasters.
-
-		Without this branch the run falls through to the classification path,
-		which decodes the flood payload into a PredictResult where nothing binds
-		and returns it: the run lists correctly, reopens as an empty card, and
-		reports nothing, with no error raised anywhere. That is the defect the
-		wind and energy branches above were each written for.
-
-		Both raster paths are rewritten to where the files actually are. What
-		was stored is relative to the data directory, and the sidecar's own
-		paths -- into a temporary work directory -- are long gone by the time a
-		run is reopened. A path that resolves to nothing is worse than none: the
-		export would offer a GeoTIFF that cannot be produced.
-	*/
-	if run.Kind == store.RunKindFlood {
-		var flood analysis.FloodAnalysis
-		if run.ResultJSON != "" && run.ResultJSON != "{}" {
-			_ = json.Unmarshal([]byte(run.ResultJSON), &flood)
-		}
-		flood.NormalizeNilSlices()
-		/*
-			Extent comes back with the rest of the payload and is not rebuilt
-			here. It is the bounds of the AOI clip alone; Grid.Bounds is the
-			buffered window the chain ran on, several times that ground, so a
-			restore that filled the missing field from the grid would place the
-			overlay stretched over the buffer -- the same misreading in pixels
-			that reporting over the window was in numbers. A run stored before
-			the field existed therefore restores with a zero extent and no
-			overlay, which the map can detect; a plausible wrong one it cannot.
-		*/
-		png := filepath.Join(assetsDir, floodAgreementPNG)
-		flood.AgreementPNG = ""
-		if uri, err := store.ReadFileDataURI(png, "image/png"); err == nil {
-			flood.AgreementURI = uri
-			flood.AgreementPNG = png
-		}
-		flood.AgreementTIF = ""
-		tif := filepath.Join(assetsDir, floodAgreementTIF)
-		if _, err := os.Stat(tif); err == nil {
-			flood.AgreementTIF = tif
-		}
-		return &analysis.PredictResult{Flood: &flood, RunID: run.ID}, nil
-	}
-
-	if run.Kind == store.RunKindWater {
-		var water analysis.WaterAnalysis
-		if run.ResultJSON != "" && run.ResultJSON != "{}" {
-			_ = json.Unmarshal([]byte(run.ResultJSON), &water)
-		}
-		if uri, err := store.ReadFileDataURI(
-			filepath.Join(assetsDir, "water_occurrence.png"), "image/png",
-		); err == nil {
-			water.OccurrenceURI = uri
-		}
-		return &analysis.PredictResult{Water: &water}, nil
-	}
-
-	var res analysis.PredictResult
-	if run.ResultJSON != "" && run.ResultJSON != "{}" {
-		_ = json.Unmarshal([]byte(run.ResultJSON), &res)
-	}
-	if uri, err := store.ReadFileDataURI(filepath.Join(assetsDir, "overlay.png"), "image/png"); err == nil {
-		res.OverlayURI = uri
-	}
-	if uri, err := store.ReadFileDataURI(filepath.Join(assetsDir, "confidence.png"), "image/png"); err == nil {
-		res.ConfidenceURI = uri
-	}
-	if uri, err := store.ReadFileDataURI(filepath.Join(assetsDir, "ndvi_mean.png"), "image/png"); err == nil {
-		res.NDVIMeanURI = uri
-	}
-	if uri, err := store.ReadFileDataURI(filepath.Join(assetsDir, "true_color.png"), "image/png"); err == nil {
-		res.TrueColorURI = uri
-	}
-	if uri, err := store.ReadFileDataURI(filepath.Join(assetsDir, "reference.png"), "image/png"); err == nil {
-		res.ReferenceURI = uri
-	}
-	if res.LULC != nil {
-		if uri, err := store.ReadFileDataURI(filepath.Join(assetsDir, "lulc_map.png"), "image/png"); err == nil {
-			res.LULC.MapURI = uri
-		}
-	} else {
-		// Older saves may lack lulc block; map alone is optional.
-		if uri, err := store.ReadFileDataURI(filepath.Join(assetsDir, "lulc_map.png"), "image/png"); err == nil {
-			res.LULC = &analysis.LULCAnalysis{MapURI: uri}
-		}
-	}
-	tif := filepath.Join(assetsDir, "classification.tif")
-	if _, err := os.Stat(tif); err == nil {
-		res.RasterTIF = tif
-	}
-	if res.DateRange == nil {
-		res.DateRange = []string{run.PeriodStart, run.PeriodEnd}
-	}
-	if res.NDates == 0 {
-		res.NDates = run.NDates
-	}
-	/*
-		The run this result IS, stamped on the way out.
-
-		RunID is written by Predict after persisting, so the copy that goes to
-		disk does not carry it -- the id did not exist when the result was
-		serialised. Reloading therefore handed back a result that could not say
-		which run it was, and every surface downstream had to guess. The map
-		screen guessed the literal "current", which the board then listed as a
-		run named after the AOI label or, failing that, "Analysis": a row that
-		looked like a saved analysis, could not be told apart from one, and
-		refused every action keyed on its id -- deleting it asked the store for
-		a run called "current" and was told, correctly, that there is none.
-
-		Set here rather than at each caller because the id is only knowable
-		here, and a caller that forgets is a caller that reintroduces the ghost.
-	*/
-	res.RunID = run.ID
-	return &res, nil
-}
 
 // GeocodeSearch resolves a place name to candidate locations (OSM Nominatim).
 func (a *App) GeocodeSearch(query string) ([]geocode.GeocodeResult, error) {
@@ -1772,202 +322,3 @@ func (a *App) OpenExternal(url string) {
 }
 
 // --- Auth / profile ---
-
-func (a *App) setSession(u *store.User, token string) {
-	a.mu.Lock()
-	defer a.mu.Unlock()
-	a.currentUser = u
-	a.sessionToken = token
-}
-
-func (a *App) clearSession() {
-	a.mu.Lock()
-	defer a.mu.Unlock()
-	a.currentUser = nil
-	a.sessionToken = ""
-}
-
-// Register creates a local account and starts a session.
-func (a *App) Register(email, password, displayName string) (*store.User, error) {
-	st, err := a.requireStore()
-	if err != nil {
-		return nil, err
-	}
-	u, token, err := st.Register(email, password, displayName)
-	if err != nil {
-		return nil, mapStoreErr(err)
-	}
-	a.setSession(u, token)
-	return u, nil
-}
-
-// Login authenticates and starts a session.
-func (a *App) Login(email, password string) (*store.User, error) {
-	st, err := a.requireStore()
-	if err != nil {
-		return nil, err
-	}
-	u, token, err := st.Login(email, password)
-	if err != nil {
-		return nil, mapStoreErr(err)
-	}
-	a.setSession(u, token)
-	return u, nil
-}
-
-// Logout ends the current session.
-func (a *App) Logout() error {
-	st := a.currentStore()
-	if st == nil {
-		a.clearSession()
-		return nil
-	}
-	a.mu.RLock()
-	token := a.sessionToken
-	a.mu.RUnlock()
-	_ = st.Logout(token)
-	a.clearSession()
-	return nil
-}
-
-// CurrentUser returns the logged-in user, or nil.
-func (a *App) CurrentUser() *store.User {
-	a.mu.RLock()
-	defer a.mu.RUnlock()
-	return a.currentUser
-}
-
-// UpdateProfile updates the display name.
-func (a *App) UpdateProfile(displayName string) (*store.User, error) {
-	st, err := a.requireStore()
-	if err != nil {
-		return nil, err
-	}
-	a.mu.RLock()
-	u := a.currentUser
-	a.mu.RUnlock()
-	if u == nil {
-		return nil, store.ErrUnauthorized
-	}
-	updated, err := st.UpdateProfile(u.ID, displayName)
-	if err != nil {
-		return nil, mapStoreErr(err)
-	}
-	a.mu.Lock()
-	a.currentUser = updated
-	a.mu.Unlock()
-	return updated, nil
-}
-
-// SetAvatar saves a profile photo from a browser data URI (data:image/...;base64,...).
-func (a *App) SetAvatar(dataURI string) (*store.User, error) {
-	st, err := a.requireStore()
-	if err != nil {
-		return nil, err
-	}
-	a.mu.RLock()
-	u := a.currentUser
-	a.mu.RUnlock()
-	if u == nil {
-		return nil, store.ErrUnauthorized
-	}
-	updated, err := st.SetAvatarFromDataURI(u.ID, dataURI)
-	if err != nil {
-		return nil, mapStoreErr(err)
-	}
-	a.mu.Lock()
-	a.currentUser = updated
-	a.mu.Unlock()
-	return updated, nil
-}
-
-// ClearAvatar removes the current user's profile photo.
-func (a *App) ClearAvatar() (*store.User, error) {
-	st, err := a.requireStore()
-	if err != nil {
-		return nil, err
-	}
-	a.mu.RLock()
-	u := a.currentUser
-	a.mu.RUnlock()
-	if u == nil {
-		return nil, store.ErrUnauthorized
-	}
-	updated, err := st.ClearAvatar(u.ID)
-	if err != nil {
-		return nil, mapStoreErr(err)
-	}
-	a.mu.Lock()
-	a.currentUser = updated
-	a.mu.Unlock()
-	return updated, nil
-}
-
-// GetPreferences returns preferences for the logged-in user (or local guest).
-func (a *App) GetPreferences() (*store.Preferences, error) {
-	st, err := a.requireStore()
-	if err != nil {
-		return nil, err
-	}
-	return st.GetPreferences(a.effectiveUserID())
-}
-
-// SavePreferences persists preferences for the logged-in user (or local guest).
-func (a *App) SavePreferences(prefs store.Preferences) error {
-	st, err := a.requireStore()
-	if err != nil {
-		return err
-	}
-	prefs.UserID = a.effectiveUserID()
-	return st.SavePreferences(prefs)
-}
-
-func mapStoreErr(err error) error {
-	switch {
-	case errors.Is(err, store.ErrEmailTaken):
-		return errors.New("email already registered")
-	case errors.Is(err, store.ErrInvalidCreds):
-		return errors.New("invalid email or password")
-	case errors.Is(err, store.ErrUnauthorized):
-		return errors.New("not authenticated")
-	case errors.Is(err, store.ErrInvalidInput):
-		return errors.New("invalid input")
-	case errors.Is(err, store.ErrNotFound):
-		return errors.New("not found")
-	default:
-		return err
-	}
-}
-
-// makeRunLabel builds run-<slug>-<yyyyMMdd-HHmmss> from an AOI hint.
-func makeRunLabel(aoiHint string) string {
-	slug := slugifyRunHint(aoiHint)
-	if slug == "" {
-		slug = "aoi"
-	}
-	stamp := time.Now().Format("20060102-150405")
-	return fmt.Sprintf("run-%s-%s", slug, stamp)
-}
-
-func slugifyRunHint(s string) string {
-	s = strings.ToLower(strings.TrimSpace(s))
-	var b strings.Builder
-	lastDash := false
-	for _, r := range s {
-		ok := (r >= 'a' && r <= 'z') || (r >= '0' && r <= '9')
-		if ok {
-			b.WriteRune(r)
-			lastDash = false
-			continue
-		}
-		if !lastDash {
-			b.WriteByte('-')
-			lastDash = true
-		}
-	}
-	out := strings.Trim(b.String(), "-")
-	if len(out) > 28 {
-		out = strings.TrimRight(out[:28], "-")
-	}
-	return out
-}
