+++ app_analysis.go
@@ -1,0 +1,337 @@
+package main
+
+import (
+	"errors"
+	"net/http"
+	"strconv"
+	"strings"
+
+	"geosense-infer/internal/analysis"
+
+	"github.com/google/uuid"
+)
+
+// The analyses the frontend asks for by name, and the mesh route one of them
+// serves. Each method reads its arguments, calls the runner, and hands back
+// what the sidecar returned; persisting what came back is app_runs.go and the
+// files it wrote are app_storage.go.
+
+// ListEmbeddedAreas returns the embedded study areas (A/B/C).
+func (a *App) ListEmbeddedAreas() []analysis.Area {
+	runner := a.currentRunner()
+	if runner == nil {
+		return []analysis.Area{}
+	}
+	return runner.ListAreas()
+}
+
+// Predict runs the inference sidecar for the given request.
+func (a *App) Predict(req analysis.PredictRequest) (*analysis.PredictResult, error) {
+	runner := a.currentRunner()
+	if runner == nil {
+		return nil, errors.New("runner not initialized")
+	}
+	res, err := runner.Predict(a.ctx, req)
+	if err != nil {
+		return nil, err
+	}
+	// Set after persisting, so the stored copy -- marshalled inside -- does not
+	// carry a run's own id inside its own row. The frontend needs it to attach
+	// compositions made while this run is on screen.
+	res.RunID = a.persistRunIfLoggedIn(req, res)
+	return res, nil
+}
+
+// AnalyzeLULC runs descriptive MapBiomas land-cover / land-use analysis
+// without Sentinel imagery. Embedded areas use local TIFFs; custom AOIs in
+// Brazil fetch a MapBiomas Collection 10 COG window on demand.
+func (a *App) AnalyzeLULC(req analysis.LULCRequest) (*analysis.LULCAnalysis, error) {
+	runner := a.currentRunner()
+	if runner == nil {
+		return nil, errors.New("runner not initialized")
+	}
+	return runner.AnalyzeLULC(a.ctx, req)
+}
+
+// ListDataCube inventories Sentinel-2 L2A scenes for the AOI (before Classify).
+func (a *App) ListDataCube(req analysis.DataCubeRequest) (*analysis.DataCubeResult, error) {
+	runner := a.currentRunner()
+	if runner == nil {
+		return nil, errors.New("runner not initialized")
+	}
+	return runner.ListDataCube(a.ctx, req)
+}
+
+// RenderComposite builds an RGB / false-color or spectral-index overlay for one scene.
+func (a *App) RenderComposite(req analysis.CompositeRequest) (*analysis.CompositeResult, error) {
+	runner := a.currentRunner()
+	if runner == nil {
+		return nil, errors.New("runner not initialized")
+	}
+	return runner.RenderComposite(a.ctx, req)
+}
+
+// AnalyzeWater maps surface water over a period from spectral water indices.
+// Descriptive: a thresholded index, with no model and no trained legend.
+func (a *App) AnalyzeWater(req analysis.WaterRequest) (*analysis.WaterAnalysis, error) {
+	runner := a.currentRunner()
+	if runner == nil {
+		return nil, errors.New("runner not initialized")
+	}
+	res, err := runner.AnalyzeWater(a.ctx, req)
+	if err != nil {
+		return nil, err
+	}
+	a.persistWaterRun(req, res)
+	return res, nil
+}
+
+/*
+savedRun is what one product contributes to a run row: the parts the writer
+below cannot know.
+
+Six products persist a run, and apart from these fields the path is one
+sequence written six times. That is how it drifted: the "run-" prefix, the
+trimmed project id and the AoiID column each had to be added in every copy,
+and the thumbnail column the classification path fills never reached any of
+the others.
+*/
+
+// AnalyzeSolar computes the solar resource and photovoltaic yield at the AOI.
+func (a *App) AnalyzeSolar(req analysis.SolarRequest) (*analysis.SolarAnalysis, error) {
+	runner := a.currentRunner()
+	if runner == nil {
+		return nil, errors.New("runner not initialized")
+	}
+	res, err := runner.AnalyzeSolar(a.ctx, req)
+	if err != nil {
+		return nil, err
+	}
+	a.persistSolarRun(req, res)
+	return res, nil
+}
+
+// AnalyzeDomainShift compares two cached domain fingerprints for shift diagnosis.
+func (a *App) AnalyzeDomainShift(req analysis.DomainShiftRequest) (*analysis.DomainShiftReport, error) {
+	runner := a.currentRunner()
+	if runner == nil {
+		return nil, errors.New("runner not initialized")
+	}
+	return runner.AnalyzeDomainShift(a.ctx, req)
+}
+
+// AnalyzeDomainShiftCohort measures one source AOI against every target at once.
+func (a *App) AnalyzeDomainShiftCohort(
+	req analysis.DomainShiftCohortRequest,
+) (*analysis.DomainShiftCohort, error) {
+	runner := a.currentRunner()
+	if runner == nil {
+		return nil, errors.New("runner not initialized")
+	}
+	return runner.AnalyzeDomainShiftCohort(a.ctx, req)
+}
+
+// BuildCanopyField returns the leaf-area-density field of one orchard module,
+// together with the transmittances the GLSL march has to reproduce.
+//
+// Not persisted as a run: the field is a function of its parameters and costs
+// under a second to rebuild, so storing it would keep a copy that the next
+// change of spacing invalidates. The analyses that do get saved are the ones
+// carrying a satellite acquisition nobody can reproduce on demand.
+func (a *App) BuildCanopyField(req analysis.CanopyFieldRequest) (*analysis.CanopyField, error) {
+	runner := a.currentRunner()
+	if runner == nil {
+		return nil, errors.New("runner not initialized")
+	}
+	return runner.BuildCanopyField(a.ctx, req)
+}
+
+// BuildCanopyFromAOI reads an AOI's own vegetation-index series as a canopy:
+// LAI by date, the Helios age that carries it, and -- given a location -- what
+// that canopy intercepts under the sun the cell actually received.
+//
+// Not persisted, for the reason the other two canopy calls give: it is a
+// function of a saved run plus a sowing, and both are already recorded.
+func (a *App) BuildCanopyFromAOI(req analysis.CanopyFromAOIRequest) (*analysis.CanopyFromAOI, error) {
+	runner := a.currentRunner()
+	if runner == nil {
+		return nil, errors.New("runner not initialized")
+	}
+	return runner.BuildCanopyFromAOI(a.ctx, req)
+}
+
+// BuildCanopyMesh grows a stand of plants and returns it as glTF, for a reader
+// who wants to see the canopy rather than a density that stands for it.
+//
+// Not persisted, for the reason BuildCanopyField gives, and for one more: the
+// stand is deterministic in its seed, so the parameters are a smaller and more
+// durable record of it than the megabytes of triangles they produce.
+func (a *App) BuildCanopyMesh(req analysis.CanopyMeshRequest) (*analysis.CanopyMesh, error) {
+	runner := a.currentRunner()
+	if runner == nil {
+		return nil, errors.New("runner not initialized")
+	}
+	mesh, err := runner.BuildCanopyMesh(a.ctx, req)
+	if err != nil {
+		return nil, err
+	}
+
+	/*
+		The bytes are held here and the reply carries a URL instead.
+
+		Returning them would put a base64 string of the whole mesh through the
+		Wails bridge, which marshals every bound result to JSON. On WKWebView
+		that is where "Maximum call stack size exceeded" is thrown -- inside the
+		bridge, before any application JavaScript runs, which is why it survived
+		being verified everywhere outside the webview.
+
+		The id changes per build so the webview cannot serve a previous stand
+		from cache, and each build is held under its own id rather than
+		replacing the last -- see the field's comment for the race that made a
+		single slot wrong.
+	*/
+	id := uuid.NewString()
+
+	a.meshMu.Lock()
+	if a.meshes == nil {
+		a.meshes = make(map[string][]byte)
+	}
+	a.meshes[id] = mesh.Data
+	a.meshOrder = append(a.meshOrder, id)
+	for len(a.meshOrder) > maxHeldMeshes {
+		delete(a.meshes, a.meshOrder[0])
+		a.meshOrder = a.meshOrder[1:]
+	}
+	a.meshMu.Unlock()
+
+	mesh.Data = nil
+	mesh.URL = meshURLPrefix + id
+	return mesh, nil
+}
+
+// The path the grown stand is served from. A prefix rather than a fixed name
+// because the id changes per build, which is what keeps the webview from
+// answering a fetch out of its cache with the previous canopy.
+const meshURLPrefix = "/canopy-mesh/"
+
+/*
+meshMiddleware serves the last grown stand as bytes.
+
+AssetServer middleware rather than its Handler, and the distinction is the whole
+reason this works: Handler is consulted only when Assets reports the file
+missing, and a single-page front end answers any unknown path with index.html
+instead. A mesh request therefore came back as HTML and the loader reported
+"Unrecognized token '<'". Middleware sits ahead of Assets, so this decides its
+own route and passes everything else through untouched.
+*/
+
+func (a *App) meshMiddleware(next http.Handler) http.Handler {
+	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		if !strings.HasPrefix(r.URL.Path, meshURLPrefix) {
+			next.ServeHTTP(w, r)
+			return
+		}
+		id := strings.TrimPrefix(r.URL.Path, meshURLPrefix)
+
+		a.meshMu.RLock()
+		data, held := a.meshes[id]
+		a.meshMu.RUnlock()
+
+		// An id nobody is holding is one that has aged out, or one that was
+		// never issued. Either way this must answer rather than fall through:
+		// the asset server behind it replies to unknown paths with index.html,
+		// and a loader handed HTML reports "Unrecognized token '<'".
+		if id == "" || !held || len(data) == 0 {
+			http.NotFound(w, r)
+			return
+		}
+		w.Header().Set("Content-Type", "model/gltf-binary")
+		w.Header().Set("Content-Length", strconv.Itoa(len(data)))
+		// The id is unique per build, so the bytes behind a URL never change.
+		w.Header().Set("Cache-Control", "no-store")
+		w.WriteHeader(http.StatusOK)
+		_, _ = w.Write(data)
+	})
+}
+
+// AnalyzeSolarTerrain maps plane-of-array irradiation over the AOI terrain.
+func (a *App) AnalyzeSolarTerrain(req analysis.SolarTerrainRequest) (*analysis.SolarTerrainAnalysis, error) {
+	runner := a.currentRunner()
+	if runner == nil {
+		return nil, errors.New("runner not initialized")
+	}
+	res, err := runner.AnalyzeSolarTerrain(a.ctx, req)
+	if err != nil {
+		return nil, err
+	}
+	a.persistSolarRaster(req.AreaID, req.PolygonGeoJSON, req.Label, req.RunLabel,
+		req.ProjectID, req.AoiID, "solar_terrain", res.Season, res, res.OverlayURI, res.NDates())
+	return res, nil
+}
+
+// AnalyzeSolarSiting classifies the AOI for fixed-tilt photovoltaic siting.
+func (a *App) AnalyzeSolarSiting(req analysis.SolarSitingRequest) (*analysis.SolarSitingAnalysis, error) {
+	runner := a.currentRunner()
+	if runner == nil {
+		return nil, errors.New("runner not initialized")
+	}
+	res, err := runner.AnalyzeSolarSiting(a.ctx, req)
+	if err != nil {
+		return nil, err
+	}
+	a.persistSolarRaster(req.AreaID, req.PolygonGeoJSON, req.Label, req.RunLabel,
+		req.ProjectID, req.AoiID, "solar_siting", "siting", res, res.OverlayURI, 0)
+	return res, nil
+}
+
+// AnalyzeEnergyModel runs the photovoltaic energy model over the AOI.
+func (a *App) AnalyzeEnergyModel(req analysis.EnergyModelRequest) (*analysis.EnergyModelAnalysis, error) {
+	runner := a.currentRunner()
+	if runner == nil {
+		return nil, errors.New("runner not initialized")
+	}
+	res, err := runner.AnalyzeEnergyModel(a.ctx, req)
+	if err != nil {
+		return nil, err
+	}
+	a.persistEnergyModelRun(req, res)
+	return res, nil
+}
+
+// AnalyzeWind screens the wind resource at the AOI.
+func (a *App) AnalyzeWind(req analysis.WindRequest) (*analysis.WindAnalysis, error) {
+	runner := a.currentRunner()
+	if runner == nil {
+		return nil, errors.New("runner not initialized")
+	}
+	res, err := runner.AnalyzeWind(a.ctx, req)
+	if err != nil {
+		return nil, err
+	}
+	a.persistWindRun(req, res)
+	return res, nil
+}
+
+func (a *App) AnalyzeFlood(req analysis.FloodRequest) (*analysis.FloodAnalysis, error) {
+	runner := a.currentRunner()
+	if runner == nil {
+		return nil, errors.New("runner not initialized")
+	}
+	res, err := runner.AnalyzeFlood(a.ctx, req)
+	if err != nil {
+		return nil, err
+	}
+	a.persistFloodRun(req, res)
+	return res, nil
+}
+
+// floodProductIDs lists which DEM products the envelope was measured over, for
+// the run row. The envelope is a property of the set, so a range listed without
+// the set it spans is not attributable to anything.
+func floodProductIDs(products []analysis.FloodProduct) []string {
+	ids := make([]string, 0, len(products))
+	for _, p := range products {
+		ids = append(ids, p.ID)
+	}
+	return ids
+}
